remove ctrl from direct optimizer python
